Refactor code, merge p6 checks modules
